+++ frontend/src/pages/Landing.tsx
@@ -1,4 +1,3 @@
-import { useState } from 'react';
 import { styled } from 'styled-components';
 
 import Footer from '@Components/common/Footer/Footer';
@@ -9,35 +8,9 @@ import MemberProfile from '@Components/landing/MemberProfile/MemberProfile';
 import SideLink from '@Components/landing/SideLink/SideLik';
 import StartSection from '@Components/landing/StartSection/StartSection';
 
-import { useNotification } from '@Contexts/NotificationProvider';
-
 const Landing = () => {
-  const { send } = useNotification();
-  const [number, setNumber] = useState(1);
-
   return (
     <>
-      <button
-        onClick={() => {
-          send({
-            type: 'error',
-            message: String(number),
-          });
-          setNumber((prev) => prev + 1);
-        }}
-      >
-        error 클릭!
-      </button>
-      <button
-        onClick={() => {
-          send({
-            type: 'success',
-            message: 'ㅋㅋㅋㅋ',
-          });
-        }}
-      >
-        success 클릭!
-      </button>
       <LandingHeader>
         <Header />
         <MemberProfile />
